# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Owner {
  id: number;
  nom: string;
  prenom: string;
  /** @nullable */
  email?: string | null;
  telephone: string;
  /** @nullable */
  adresse?: string | null;
  createdAt: string;
}

export interface CreateOwnerBody {
  nom: string;
  prenom: string;
  /** @nullable */
  email?: string | null;
  telephone: string;
  /** @nullable */
  adresse?: string | null;
}

export interface Patient {
  id: number;
  nom: string;
  espece: string;
  /** @nullable */
  race?: string | null;
  sexe: string;
  /** @nullable */
  dateNaissance?: string | null;
  /** @nullable */
  poids?: number | null;
  /** @nullable */
  couleur?: string | null;
  sterilise: boolean;
  ownerId: number;
  /** @nullable */
  antecedents?: string | null;
  /** @nullable */
  allergies?: string | null;
  createdAt: string;
}

export type PatientWithOwner = Patient & {
  owner?: Owner;
};

export interface CreatePatientBody {
  nom: string;
  espece: string;
  /** @nullable */
  race?: string | null;
  sexe: string;
  /** @nullable */
  dateNaissance?: string | null;
  /** @nullable */
  poids?: number | null;
  /** @nullable */
  couleur?: string | null;
  sterilise: boolean;
  ownerId: number;
  /** @nullable */
  antecedents?: string | null;
  /** @nullable */
  allergies?: string | null;
}

export interface Consultation {
  id: number;
  patientId: number;
  veterinaire: string;
  date: string;
  statut: string;
  /** @nullable */
  motif?: string | null;
  /** @nullable */
  anamnese?: string | null;
  /** @nullable */
  examenClinique?: string | null;
  /** @nullable */
  examensComplementaires?: string | null;
  /** @nullable */
  diagnostic?: string | null;
  /** @nullable */
  diagnosticIA?: string | null;
  /** @nullable */
  ordonnance?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  poids?: number | null;
  /** @nullable */
  temperature?: number | null;
  createdAt: string;
}

export type ConsultationWithPatient = Consultation & {
  patient?: PatientWithOwner;
};

export interface Acte {
  id: number;
  code: string;
  nom: string;
  categorie: string;
  prixDefaut: number;
  /** @nullable */
  description?: string | null;
  unite: string;
}

export interface ActeConsultation {
  id: number;
  acteId: number;
  consultationId: number;
  quantite: number;
  prixUnitaire: number;
  /** @nullable */
  description?: string | null;
  acte?: Acte;
}

export interface Facture {
  id: number;
  consultationId: number;
  numero: string;
  montantHT: number;
  tva: number;
  montantTTC: number;
  statut: string;
  dateEmission: string;
  /** @nullable */
  datePaiement?: string | null;
  createdAt: string;
}

export type ConsultationFull = ConsultationWithPatient & {
  actes?: ActeConsultation[];
  facture?: Facture;
};

export interface CreateConsultationBody {
  patientId: number;
  veterinaire: string;
  date: string;
  /** @nullable */
  motif?: string | null;
  statut: string;
}

export interface ActeConsultationInput {
  acteId: number;
  quantite: number;
  prixUnitaire: number;
  /** @nullable */
  description?: string | null;
}

export interface UpdateConsultationBody {
  statut?: string;
  /** @nullable */
  motif?: string | null;
  /** @nullable */
  anamnese?: string | null;
  /** @nullable */
  examenClinique?: string | null;
  /** @nullable */
  examensComplementaires?: string | null;
  /** @nullable */
  diagnostic?: string | null;
  /** @nullable */
  diagnosticIA?: string | null;
  /** @nullable */
  ordonnance?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  poids?: number | null;
  /** @nullable */
  temperature?: number | null;
  actes?: ActeConsultationInput[];
}

export interface CreateActeBody {
  code: string;
  nom: string;
  categorie: string;
  prixDefaut: number;
  /** @nullable */
  description?: string | null;
  unite: string;
}

export type FactureWithDetails = Facture & {
  consultation?: ConsultationWithPatient;
};

export interface UpdateFactureStatutBody {
  statut: string;
  /** @nullable */
  datePaiement?: string | null;
}

export interface OrdonnanceResponse {
  ordonnance: string;
}

export interface DiagnosticIARequest {
  espece: string;
  /** @nullable */
  race?: string | null;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  poids?: number | null;
  sexe: string;
  sterilise: boolean;
  anamnese: string;
  examenClinique: string;
  /** @nullable */
  examensComplementaires?: string | null;
  /** @nullable */
  antecedents?: string | null;
  /** @nullable */
  allergies?: string | null;
}

export interface DiagnosticItem {
  nom: string;
  probabilite: string;
  description: string;
}

export interface DiagnosticIAResponse {
  diagnostics: DiagnosticItem[];
  recommandations: string;
  urgence: string;
  texteComplet: string;
}

export interface DashboardStats {
  totalPatients: number;
  totalConsultations: number;
  consultationsAujourdhui: number;
  consultationsEnCours: number;
  chiffreAffaireMois: number;
  facturesImpayees: number;
  totalProprietaires: number;
}

export type ListOwnersParams = {
  search?: string;
};

export type ListPatientsParams = {
  search?: string;
  ownerId?: number;
  espece?: string;
};

export type ListConsultationsParams = {
  statut?: string;
  date?: string;
};

export type ListFacturesParams = {
  statut?: string;
};