# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetCare Pro API - Logiciel vétérinaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}/consultations:
    get:
      operationId: listPatientConsultations
      tags: [patients]
      summary: Historique des consultations d'un patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultations du patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: Liste des consultations
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationWithPatient"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Créer une consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationBody"
      responses:
        "201":
          description: Consultation créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Détails d'une consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationFull"
        "404":
          description: Non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Mettre à jour une consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationBody"
      responses:
        "200":
          description: Consultation mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}/ordonnance:
    post:
      operationId: generateOrdonnance
      tags: [consultations]
      summary: Générer une ordonnance automatique avec IA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordonnance générée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdonnanceResponse"
  /consultations/{id}/facture:
    post:
      operationId: generateFacture
      tags: [consultations]
      summary: Générer la facture d'une consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facture générée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
  /actes:
    get:
      operationId: listActes
      tags: [actes]
      summary: Liste des actes médicaux
      responses:
        "200":
          description: Liste des actes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Acte"
    post:
      operationId: createActe
      tags: [actes]
      summary: Créer un acte médical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActeBody"
      responses:
        "201":
          description: Acte créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acte"
  /actes/{id}:
    patch:
      operationId: updateActe
      tags: [actes]
      summary: Mettre à jour un acte
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActeBody"
      responses:
        "200":
          description: Acte mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acte"
    delete:
      operationId: deleteActe
      tags: [actes]
      summary: Supprimer un acte
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Supprimé
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Owner:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        email:
          type: ["string", "null"]
        telephone:
          type: string
        adresse:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - prenom
        - telephone
        - createdAt
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        espece:
          type: string
        race:
          type: ["string", "null"]
        sexe:
          type: string
        dateNaissance:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        couleur:
          type: ["string", "null"]
        sterilise:
          type: boolean
        ownerId:
          type: integer
        antecedents:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - nom
        - espece
        - sexe
        - sterilise
        - ownerId
        - createdAt
    PatientWithOwner:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            owner:
              $ref: "#/components/schemas/Owner"
...
    Consultation:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        veterinaire:
          type: string
        date:
          type: string
        statut:
          type: string
        motif:
          type: ["string", "null"]
        anamnese:
          type: ["string", "null"]
        examenClinique:
          type: ["string", "null"]
        examensComplementaires:
          type: ["string", "null"]
        diagnostic:
          type: ["string", "null"]
        diagnosticIA:
          type: ["string", "null"]
        ordonnance:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - veterinaire
        - date
        - statut
        - createdAt
    ConsultationWithPatient:
      allOf:
        - $ref: "#/components/schemas/Consultation"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/PatientWithOwner"
    ConsultationFull:
      allOf:
        - $ref: "#/components/schemas/ConsultationWithPatient"
        - type: object
          properties:
            actes:
              type: array
              items:
                $ref: "#/components/schemas/ActeConsultation"
            facture:
              $ref: "#/components/schemas/Facture"
    CreateConsultationBody:
      type: object
      properties:
        patientId:
          type: integer
        veterinaire:
          type: string
        date:
          type: string
        motif:
          type: ["string", "null"]
        statut:
          type: string
      required:
        - patientId
        - veterinaire
        - date
        - statut
    UpdateConsultationBody:
      type: object
      properties:
        statut:
          type: string
        motif:
          type: ["string", "null"]
        anamnese:
          type: ["string", "null"]
        examenClinique:
          type: ["string", "null"]
        examensComplementaires:
          type: ["string", "null"]
        diagnostic:
          type: ["string", "null"]
        diagnosticIA:
          type: ["string", "null"]
        ordonnance:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        poids:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        actes:
          type: array
          items:
            $ref: "#/components/schemas/ActeConsultationInput"
    ActeConsultation:
      type: object
      properties:
        id:
          type: integer
        acteId:
          type: integer
        consultationId:
          type: integer
        quantite:
          type: integer
        prixUnitaire:
          type: number
        description:
          type: ["string", "null"]
        acte:
          $ref: "#/components/schemas/Acte"
      required:
        - id
        - acteId
        - consultationId
        - quantite
        - prixUnitaire
    ActeConsultationInput:
      type: object
      properties:
        acteId:
          type: integer
        quantite:
          type: integer
        prixUnitaire:
          type: number
        description:
          type: ["string", "null"]
      required:
        - acteId
        - quantite
        - prixUnitaire
    Acte:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        nom:
          type: string
        categorie:
          type: string
        prixDefaut:
          type: number
        description:
          type: ["string", "null"]
        unite:
          type: string
      required:
        - id
        - code
        - nom
        - categorie
        - prixDefaut
        - unite
    CreateActeBody:
      type: object
      properties:
        code:
          type: string
        nom:
          type: string
        categorie:
          type: string
        prixDefaut:
          type: number
        description:
          type: ["string", "null"]
        unite:
          type: string
      required:
        - code
        - nom
        - categorie
        - prixDefaut
        - unite
    Facture:
      type: object
      properties:
        id:
          type: integer
        consultationId:
          type: integer
        numero:
          type: string
        montantHT:
          type: number
        tva:
          type: number
        montantTTC:
          type: number
        statut:
          type: string
        dateEmission:
          type: string
        datePaiement:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - consultationId
        - numero
        - montantHT
        - tva
        - montantTTC
        - statut
        - dateEmission
        - createdAt
...
    OrdonnanceResponse:
      type: object
      properties:
        ordonnance:
          type: string
      required:
        - ordonnance
...