# lib/api-zod/src/generated/types/acte.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */

export interface Acte {
  id: number;
  code: string;
  nom: string;
  categorie: string;
  prixDefaut: number;
  /** @nullable */
  description?: string | null;
  unite: string;
}