# lib/api-zod/src/generated/types/acteConsultationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */

export interface ActeConsultationInput {
  acteId: number;
  quantite: number;
  prixUnitaire: number;
  /** @nullable */
  description?: string | null;
}