# lib/api-zod/src/generated/types/consultationFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */
import type { ActeConsultation } from "./acteConsultation";
import type { ConsultationWithPatient } from "./consultationWithPatient";
import type { Facture } from "./facture";

export type ConsultationFull = ConsultationWithPatient & {
  actes?: ActeConsultation[];
  facture?: Facture;
};