# lib/api-zod/src/generated/types/consultationWithPatient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */
import type { Consultation } from "./consultation";
import type { PatientWithOwner } from "./patientWithOwner";

export type ConsultationWithPatient = Consultation & {
  patient?: PatientWithOwner;
};