# lib/api-zod/src/generated/types/ordonnanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */

export interface OrdonnanceResponse {
  ordonnance: string;
}