# lib/api-zod/src/generated/types/owner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */

export interface Owner {
  id: number;
  nom: string;
  prenom: string;
  /** @nullable */
  email?: string | null;
  telephone: string;
  /** @nullable */
  adresse?: string | null;
  createdAt: string;
}