# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VetCare Pro API - Logiciel vétérinaire
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  nom: string;
  espece: string;
  /** @nullable */
  race?: string | null;
  sexe: string;
  /** @nullable */
  dateNaissance?: string | null;
  /** @nullable */
  poids?: number | null;
  /** @nullable */
  couleur?: string | null;
  sterilise: boolean;
  ownerId: number;
  /** @nullable */
  antecedents?: string | null;
  /** @nullable */
  allergies?: string | null;
  createdAt: string;
}